+++ CHANGELOG.md
@@ -17,6 +17,7 @@ BUG FIXES:
   - guests/fedora: Fix hostname setting. [GH-3382]
   - guests/fedora: Support predictable network interface names for
       public/private networks. [GH-3207]
+  - providers/hyperv: Check for PowerShell features. [GH-3398]
   - synced\_folders/smb: Only set the chmod properly by default on Windows
       if it isn't already set. [GH-3394]
   - synced\_folders/smb: Sharing folders with odd characters like parens